Add refresh functionality (#19)
* clean: remove dummy posts

* feat(timeline): add refresh button functionality

* clean: remove console.log calls

* fix: fix posts from being sorted in reverse chronological order
