+++ src/lib/components/ui/navbar.svelte
@@ -18,6 +18,7 @@
     }
 
     export let userInfo: object = {};
+    export let refreshFunction;
 </script>
 
 <nav class="sticky top-0 left-0 right-0 bg-white text-white p-4 border-b-2 border-solid border-gray-200 max-w-screen">
@@ -134,7 +135,7 @@
         <img src="/kaleido-logomark.svg" alt="Kaleido" class="w-12 h-12" />
         
         <!-- Refresh -->
-        <Button class="bg-kaleido w-10 h-10 rounded-full p-2 border-2 border-transparent hover:bg-transparent hover:border-kaleido hover:text-kaleido">
+        <Button on:click={refreshFunction} class="bg-kaleido w-10 h-10 rounded-full p-2 border-2 border-transparent hover:bg-transparent hover:border-kaleido hover:text-kaleido">
             <RefreshCw class="w-full h-full" strokeWidth={2}/>
         </Button>
     </div>
