+++ src/routes/timeline/+layout.server.ts
@@ -1,0 +1,9 @@
+import type { PageServerLoad } from "../$types";
+
+export const load: PageServerLoad = async ({ cookies }) => {
+    return {
+        bskyToken: cookies.get("bskyToken"),
+        mastodonToken: cookies.get("mastodonToken"),
+        mastodonInstance: cookies.get("mastodonInstance"),
+    }
+};
