+++ src/routes/timeline/+layout.svelte
@@ -1,7 +1,8 @@
 <script>
     import Navbar from "$lib/components/ui/navbar.svelte";
+    import { refreshing, refreshTimeline, timelineData } from "./timelineData";
     export let data;
-    console.log(data);
+
     const userInfo = {
         mastodonPicture: data.mastodonPicture,
         mastodonHandle: data.mastodonUsername,
@@ -13,11 +14,22 @@
         bskyDisplayName: data.bskyDisplayName,
         bskyPicture: data.bskyPicture,
     }
+    const catchRefreshClick = async () => {
+        //temporarily set the timeline to be empty
+        timelineData.set({
+            mastodonTimeline: [],
+            bskyTimeline: [],
+            timeline: [],
+        })
+        refreshing.set(true)
+        await timelineData.set(await refreshTimeline(data.mastodonToken, data.mastodonInstance, data.bskyToken))
+        refreshing.set(false)
+    }
 </script>
 
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
 <main>
-    <Navbar userInfo={userInfo}/>
-    <slot />
+    <Navbar userInfo={userInfo} refreshFunction={catchRefreshClick}/>
+    <slot/>
 </main>
